requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Pillow==11.0.0
python-dotenv==1.0.1
django-cors-headers==4.3.1
django-storages==1.14.2
boto3==1.34.34 